+++ src/github.rs
@@ -1,1 +1,6 @@
+//! GitHub actions and workflow model.
+//!
+//! Provides serde based structs for GitHub's actions and workflow YAML.
+//!
 pub mod action;
+pub mod workflow;
